requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
pypdf
transformers
torch
huggingface-hub
